Restore tox.ini configuration, restructure and add configuration subsystem unit tests
