+++ tests/unit/configuration_subsystem/defaults.py
@@ -1,0 +1,7 @@
+""" defaults for adjacent tests
+"""
+import os
+
+from ...defaults import FIXTURES_DIR
+
+TEST_FIXTURE_DIR = os.path.join(FIXTURES_DIR, "unit", "configuration_subsystem")
