+++ tests/unit/configuration_subsystem/test_entries_sanity.py
