+++ tests/unit/configuration_subsystem/test_previous_cli.py
