+++ tox.ini
@@ -10,6 +10,8 @@ skip_missing_interpreters = true
 
 
 [testenv]
+# the pytest command line is not in the project.toml because of issues 
+# with catching breakpoints while debugging unit tests
 allowlist_externals = cat
                       rm
                       grep
@@ -17,7 +19,7 @@ deps = -r{toxinidir}/requirements.txt
        -r{toxinidir}/test-requirements.txt
 commands =
     /bin/bash -c 'grep -FInrq "UPDATE_FIXTURES = True" ./tests && exit 1 || exit 0'
-    py.test
+    py.test -vvvv --durations=100 --cov ansible_navigator --cov share --cov-report term-missing --cov-branch
 setenv =
     TERM = xterm
 passenv = HOME
